task 6

+++ 0x00-hello_world/6-size.c
@@ -10,6 +10,6 @@ int main(void)
 	printf("Size of a char: %lu.\n", (unsigned long)sizeof(char));
 	printf("Size of an int: %lu.\n", (unsigned long)sizeof(int));
 	printf("Size of a long int: %lu.\n", (unsigned long)sizeof(long int));
-	printf("Size of a long long int: %lu.\n", (unsigned long)sizeof(long));
+	printf("Size of a long long int: %lu.\n", (unsigned long)sizeof(long long));
 	printf("Size of a float: %lu.\n", (unsigned long)sizeof(float));
 }
